backend/requirements.txt
flask
flask-cors
psycopg2-binary
fastapi
uvicorn
psycopg2-binary
pydantic